requirements.txt
Django==2.0.5
djangorestframework==3.8.2
gunicorn==19.8.1
PyYAML==3.12
requests==2.18.4
